requirements.txt
streamlit
google-cloud-aiplatform
google-generativeai
pillow